Move to ORMLite
Writing new queries for the DB was becoming a pain for all the reasons that ORM frameworks exists. I didn't want to move away from SQLite because of the energy it saves due to the lack of orchestration I get with no separate DB process, so I picked ORMLite, which has nice SQLite integration. I specifically wanted something that could create the tables where necessary and AFAICT, Hibernate's community dialect for SQLite can't do so, so that option was out.

This commit doesn't contain any functional changes; it's just moved what already existed over to ORMLite (and reduced the amount of code significantly in the process).