Update schema to include households
Grouping accounts into households will allow more users on a single
instance of GateKey without all users seeing the keys of all other
users. For now, this just updates the Schema and adds a migration to
move to the new models.

The next step will be to allow admins to add households, place users in
households when registering them, and limit the visibility of keys to
users in the households to which the keys are associated.