Add household support
This commit adds support for separate households within GateKey. When
users are added, they're placed in a specific household and can only
create and list keys for that household. New users are now added with
the "RESIDENT" permissions bundle and only the initial user can add new
users and households.